Split filebased into a helper class, and seperate image and sound files
